requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.2
requests==2.32.3
gunicorn==21.2.0
PyJWT==2.8.0
Werkzeug==3.0.3
Pillow==10.4.0
openai==1.51.0
google-cloud-storage==2.18.2

